refactor(contact): refactors contact section
refactors the contact section file and restructures it into a new folder that includes the style sheet
