--- src/components/ScrollUp/styles 2.js
@@ -1,15 +1,0 @@
-import styled from "styled-components";
-
-export const UpArrowContainer = styled.div`
-  height: 100px;
-  width: 82vw;
-  display: flex;
-  justify-content: end;
-  align-content: flex-end;
-  margin-bottom: 20px;
-
-  img {
-    height: 40px;
-    cursor: pointer;
-  }
-`;
